+++ src/modules/helper/getListOfDigitallyNotAvailableBooks.ts
